# pkg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a basic ticket service written in go.",
        "title": "Go-Ticket API",
        "contact": {
            "name": "Max Greß",
            "url": "https://github.com/mgr1054/go-ticket",
            "email": "max.gress@student.reutlingen-university.de"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/secured/tickets/events/{id}": {
            "get": {
                "description": "Gives back a number of all sold tickets for this event\nallowed: admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Get Tickets By EventID",
                "operationId": "get-tickets-by-event-id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Event"
                        }
                    },
                    "401": {
                        "description": "{\"error\":\"Unauthorized for this route\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"error\": \"Tickets not found\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/secured/tickets/user/{id}": {
            "get": {
                "description": "Gives back all tickets for user\nallowed: admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Get Tickets By UserID",
                "operationId": "get-tickets-by-user-id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Ticket"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/secured/tickets/{id}": {
            "get": {
                "description": "Creates Ticket for EventID, also checks if enough capacity is available\nallowed: user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Create Ticket by EventID",
                "operationId": "create-ticket",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Ticket"
                        }
                    },
                    "401": {
                        "description": "{\"error\":\"Unauthorized for this route\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"error\": \"User not found\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"error\": \"Could not create Ticket\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes Ticket by Ticket ID, available up until one week before the event\nallowed: admin, user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Delete Ticket By ID",
                "operationId": "delete-tickets-by-user-id",
                "responses": {
                    "200": {
                        "description": "{\"message\": \"Ticket deleted\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"error\":\"Unauthorized for this route\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"error\": \"Tickets not found\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"error\": \"Could not parse time\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/token": {
            "post": {
                "description": "Generates JWT Token based on given context, checks if username and password match\nEncode JWT with username, email and role\nallowed: unsecured",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Generate Token",
                "operationId": "generate-token",
                "parameters": [
                    {
                        "description": "Create Token",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.TokenRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"error\": \"Could not create Token\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"error\": \"Password incorrect",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"error\":\"User not found\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"error\":\"Could not create Token\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.TokenRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "test@online.de"
                },
                "password": {
                    "type": "string",
                    "example": "1234"
                }
            }
        },
...
        "models.Event": {
            "type": "object",
            "properties": {
                "band_name": {
                    "type": "string"
                },
                "capacity": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                }
            }
        },
        "models.Ticket": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}